+++ Readme.md
@@ -10,9 +10,9 @@ The configuration file is appsettings.json
 |-------|-----------|
 |SpotifyApiToken|Create youself a token here https://developer.spotify.com/|
 |SpotifyUsername|Spotify username, regular or email|
-|BbcPlaylistUrls|E.G. https://www.bbc.co.uk/programmes/articles/5JDPyPdDGs3yCLdtPhGgWM7/bbc-radio-6-music-playlist|
+|WebPlaylistUrls|E.G. https://www.bbc.co.uk/programmes/articles/5JDPyPdDGs3yCLdtPhGgWM7/bbc-radio-6-music-playlist|
 
 ## Step 2 - Run
 
-Run BbcPlaylistToSpotify.exe
+Run WebPlaylistToSpotify.exe
