+++ WebPlaylistToSpotify.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32922.545
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BbcPlaylistToSpotify", "BbcPlaylistToSpotify\BbcPlaylistToSpotify.csproj", "{ED9F4F2B-88C1-4F19-83D7-44D562375F40}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebPlaylistToSpotify", "WebPlaylistToSpotify\WebPlaylistToSpotify.csproj", "{ED9F4F2B-88C1-4F19-83D7-44D562375F40}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
