+++ WebPlaylistToSpotify/appsettings.json
@@ -1,7 +1,7 @@
 {
   "SpotifyUsername": "",
   "SpotifyApiToken": "",
-  "BbcPlaylistUrls": [
+  "WebPlaylistUrls": [
     "https://www.bbc.co.uk/programmes/articles/5JDPyPdDGs3yCLdtPhGgWM7/bbc-radio-6-music-playlist",
     "https://www.bbc.co.uk/programmes/articles/3tqPdBWF9yMbTrfjWvfKV8t/radio-1-playlist"
   ]
